Move constants.js to _utils to remove from test coverage
